+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Dev: Add moderation event tests. (#21)
-- Dev: Add bits event tests. (#22)
+- Dev: Add bits event tests. (#22, #23)
 
 ## v0.1.0